refactor : svg 디렉토리 정리, import 경로 수정 노가다
