+++ src/components/kid/create/SelectItemNameButton.tsx
@@ -1,6 +1,6 @@
 import { HTMLAttributes, useState } from 'react';
 import styled from 'styled-components';
-import { ReactComponent as SelectItemNameButtonBorder } from '@assets/icons/itemName-border.svg';
+import { ReactComponent as SelectItemNameButtonBorder } from '@assets/border/itemName-border.svg';
 import { theme } from '@lib/styles/theme';
 
 interface SelectItemNameButtonProps extends HTMLAttributes<HTMLButtonElement> {
