+++ src/components/onboarding/ProfileButton.tsx
@@ -1,7 +1,7 @@
 import { HTMLAttributes, useState } from 'react';
 import styled from 'styled-components';
-import { ReactComponent as ProfileButtonCharacter } from '@assets/icons/profile-button-character.svg';
-import { ReactComponent as ProfileButtonBorder } from '@assets/icons/profile-button-border.svg';
+import { ReactComponent as ProfileButtonCharacter } from '@assets/illust/profile-button-character.svg';
+import { ReactComponent as ProfileButtonBorder } from '@assets/border/profile-button-border.svg';
 import { clacRatio, theme } from '@lib/styles/theme';
 
 interface ProfileButtonProps extends HTMLAttributes<HTMLButtonElement> {
